requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.12.5
distlib==0.3.1
Django==3.1.7
django-autoslug==1.9.8
django-crispy-forms==1.11.1
django-cuser==2017.3.16
django-modeltranslation==0.16.2
django-phonenumber-field==5.0.0
django-quill-editor==0.1.19
django-tinymce==3.2.0
filelock==3.0.12
phonenumbers==8.12.18
Pillow==8.1.0
pipenv==2020.11.15
pkg-resources==0.0.0
psycopg2-binary==2.8.6
python-dotenv==0.15.0
pytz==2021.1
selenium==3.141.0
six==1.15.0
sqlparse==0.4.1
Unidecode==1.2.0
urllib3==1.26.3
virtualenv==20.4.2
virtualenv-clone==0.5.4
